dockerize
Moved the application over to docker.

The current build process works as follows.

The frontend and backend build independently, then the web container exposes them. The frontend builds to the `frontend_build` volume and web access it.

The backend is reversed proxied by the web container.

The configuration for the web container may be put in ./web/Caddyfile. An example is located at ./web/Caddyfile.example